+++ lib/key/key.ts
@@ -1,1 +1,0 @@
-export { Key } from "@eu-ge-ne/kitty-keys";